build(deps): Adjust to OIIO changes to TextureOpt structure (#1888)
OIIO 3.0 changes the TextureOpt structure a bit, and we need to
adjust on the OSL end, too.

+++ src/liboslexec/oslexec_pvt.h
@@ -25,9 +25,12 @@
 #    include "string_hash.h"
 #endif
 
+#include <OpenImageIO/Imath.h>
+
 #include <OpenImageIO/color.h>
 #include <OpenImageIO/paramlist.h>
 #include <OpenImageIO/refcnt.h>
+#include <OpenImageIO/texture.h>
 #include <OpenImageIO/thread.h>
 #include <OpenImageIO/ustring.h>
 
@@ -2512,13 +2515,13 @@ tex_interp_to_code(StringParam modename)
 {
     int mode = -1;
     if (modename == STRING_PARAMS(smartcubic))
-        mode = TextureOpt::InterpSmartBicubic;
+        mode = (int)TextureOpt::InterpSmartBicubic;
     else if (modename == STRING_PARAMS(linear))
-        mode = TextureOpt::InterpBilinear;
+        mode = (int)TextureOpt::InterpBilinear;
     else if (modename == STRING_PARAMS(cubic))
-        mode = TextureOpt::InterpBicubic;
+        mode = (int)TextureOpt::InterpBicubic;
     else if (modename == STRING_PARAMS(closest))
-        mode = TextureOpt::InterpClosest;
+        mode = (int)TextureOpt::InterpClosest;
     return mode;
 }
